requirements.txt
# Alkaid Self-Optimization System v2
# Note: Most dependencies are already installed in Bot venvs
# This file documents what the system uses

# Already in kanban-kiro-bot/venv:
discord.py>=2.7.0
aiohttp>=3.13.0
aiosqlite>=0.22.0
python-dotenv>=1.0.0

# Already on system:
# sqlite3 (built-in)
# qmd 1.0.7 (/usr/bin/qmd)
# kiro-cli 1.27.1 (/root/.local/bin/kiro-cli)

# No additional packages needed
